refactor: move item and list-box extensions to separate files
